Revert "refactor - all code folder names changed"
