+++ src/component/custom_input/custom-input-text-style.js
