+++ src/component/home-screen-component/footerHomeScreenComponent.js
