+++ src/containers/sign-up-screen/sign-up-screen.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { connect } from "react-redux";
 import { Button } from 'react-native-elements';
 import validate from '../../utility/validation';
-import styles from './SignUpScreenStyle';
-import CustomInputText from '../../component/CustomInputText/CustomInputText';
+import styles from './sign-up-screen-style';
+import CustomInputText from '../../component/custom_input/custom-input-text';
 import { View, Text, Picker, KeyboardAvoidingView, ScrollView } from 'react-native';
 
 import {
