+++ src/navigation/appStack.js
@@ -1,8 +1,8 @@
 
 import { createStackNavigator } from 'react-navigation';
-import LoginScreen from '../screens/LoginScreen/LoginScreen';
-import HomeScreen from '../screens/HomeScreen/HomeScreen';
-import SignUpScreen from '../screens/SignUpScreen/SignUpScreen';
+import LoginScreen from '../containers/login-screen/login-screen';
+import HomeScreen from '../containers/home-screen/home-screen';
+import SignUpScreen from '../containers/sign-up-screen/sign-up-screen';
 
 const AppStack = createStackNavigator({
   Home: HomeScreen,
