+++ src/store/navigator/navigator-action/navigatorActionTypes.js
