+++ src/store/navigator/navigator-reducer/navigatorReducerData.js
